# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'My API'
  version: '1.0'
paths:
  /api/register:
    post:
      tags:
        - Authentication
      summary: 'Đăng ký'
      description: 'Đăng ký tài khoản'
      operationId: register
      requestBody:
        description: 'Nhập thông tin đăng ký tài khoản'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
                - password_confirmation
              properties:
                name:
                  description: 'User Name'
                  type: string
                  example: 'Anh Nhi'
                email:
                  description: 'User Email'
                  type: string
                  example: nhi@gmail.com
                password:
                  description: 'User password'
                  type: string
                  example: '123456'
                password_confirmation:
                  description: 'User confirm password'
                  type: string
                  example: '123456'
              type: object
      responses:
        '200':
          description: 'Đăng ký tài khoản thành công'
        '400':
          description: 'Dữ liệu không hợp lệ'
  /api/login:
    post:
      tags:
        - Authentication
      summary: 'Đăng nhập'
      description: 'Đăng nhập vào cửa hàng'
      operationId: login
      requestBody:
        description: 'Nhập thông tin đăng nhập'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  description: 'Địa chỉ email của người dùng'
                  type: string
                password:
                  description: 'Mật khẩu của người dùng'
                  type: string
              type: object
      responses:
        '200':
          description: 'Đăng nhập thành công'
        '400':
          description: 'Dữ liệu không hợp lệ'
  /api/profile:
    get:
      tags:
        - Authentication
      summary: 'Hồ sơ người dùng'
      description: 'Hồ sơ người mua hàng (người dùng)'
      operationId: show
      responses:
        '200':
          description: 'Hiển thị hồ sơ người dùng'
        '401':
          description: 'Không xác thực được người dùng'
      security:
        -
          bearer: []
  /api/logout:
    post:
      tags:
        - Authentication
      summary: 'Đăng xuất'
      description: 'Đăng xuất người dùng'
      operationId: logout
      responses:
        '200':
          description: 'Đăng xuất thành công'
        '401':
          description: 'Không xác thực được người dùng'
      security:
        -
          bearer: []
  /api/cart/add:
    post:
      tags:
        - Cart
      summary: 'Thêm vào giỏ hàng'
      description: 'Thêm sản phẩm vào giỏ hàng'
      operationId: 'App\Application\Controllers\CartController::addToCart'
      requestBody:
        description: 'Trò chơi được thêm vào giỏ hàng'
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  description: 'ID của trò chơi'
                  type: integer
              type: object
      responses:
        '200':
          description: 'Thêm vào giỏ hàng thành công'
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  /api/cart/update:
    put:
      tags:
        - Cart
      summary: 'Cập nhật giỏ hàng với ID trò chơi'
      description: 'Cập nhật sản phẩm trong giỏ hàng'
      operationId: 'App\Application\Controllers\CartController::updateCart'
      requestBody:
        description: 'Thông tin cập nhật trò chơi trong giỏ hàng'
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  description: 'ID của trò chơi'
                  type: integer
                quantity:
                  description: 'Số lượng'
                  type: integer
              type: object
      responses:
        '200':
          description: 'Đã cập nhật số lượng'
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  /api/cart:
    get:
      tags:
        - Cart
      summary: 'Hiển thị giỏ hàng'
      description: 'Hiển thị giỏ hàng hiện tại'
      operationId: 'App\Application\Controllers\CartController::showCart'
      responses:
        '200':
          description: 'Thông tin giỏ hàng hiện tại'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id: { description: 'ID của trò chơi', type: integer }, name: { description: 'Tên của trò chơi', type: string }, price: { description: Giá, type: number }, image: { description: 'Hình ảnh', type: string }, quantity: { description: 'Số lượng', type: integer }, total_price: { description: 'Tổng giá khi đã nhân với số lượng', type: number } }
                  type: object
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  '/api/cart/remove/{productId}':
    delete:
      tags:
        - Cart
      summary: 'Xóa trò chơi khỏi giỏ hàng (Xóa hết & xóa theo ID)'
      description: 'Xóa sản phẩm khỏi giỏ hàng'
      operationId: 'App\Application\Controllers\CartController::removeFromCart'
      parameters:
        -
          name: id
          in: path
          description: 'ID của trò chơi cần xóa'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 'Trò chơi đã được xóa khỏi giỏ hàng'
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
  /api/cart/buy:
    post:
      tags:
        - Cart
      summary: 'Đặt hàng & thanh toán'
      description: 'Thanh toán giỏ hàng'
      operationId: 'App\Application\Controllers\CartController::payCart'
      responses:
        '200':
          description: 'Giỏ hàng đã được thanh toán'
        '422':
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header